django-backend/requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dateparser==1.1.4
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
domaintools-api==1.0.1
elastic-apm==6.13.2
filelock==3.8.0
grpcio==1.51.1
grpcio-tools==1.51.1
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.2
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
Markdown==3.4.1
multidict==6.0.3
pipenv==2022.10.12
platformdirs==2.5.2
protobuf==4.21.12
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
sentry-sdk==1.11.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
virtualenv==20.16.5
virtualenv-clone==0.5.7
wrapt==1.14.1
